requirements.txt
functions-framework
line-bot-sdk==3.13.0
pyyaml
pyjwt
jwcrypto